feat: renamge raw config to pre-compiled for more descriptiveness, made split of config types more apparent
