+++ src/pygeoml1000/compilation.py
@@ -1,0 +1,301 @@
+"""Compilation of the pre-compiled configs into the objects the geometry is built from.
+
+The pre-compiled configs describe the shape of the setup: how many strings, how
+many detectors on a string, where the PMTs sit. This module turns them into the
+two compiled objects, ``channelmap`` and ``special_metadata``, which name every
+channel one by one.
+
+:func:`pygeoml1000.config.resolve_config` runs this step. A config that already
+holds both compiled objects skips it.
+"""
+
+from __future__ import annotations
+
+import copy
+import logging
+
+import numpy as np
+
+from .utils import convert_to_plain_types, normalize_records
+
+log = logging.getLogger(__name__)
+
+
+def calculate_and_place_pmts(channelmap: dict, configs: dict, rawid_start: int = 6000) -> None:
+    from . import watertank  # noqa: PLC0415
+
+    # Floor PMTs are pretty trivial to place
+    rawid = rawid_start
+    for row in configs["pmts_pos"]["floor"].values():
+        row_index = row["id"]
+        pmts_in_row = row["n"]
+        radius = row["r"]
+
+        for i in range(pmts_in_row):
+            name = f"PMT0{row_index}{i + 1:02d}"
+            x = radius * np.cos(np.radians(360 / pmts_in_row * i))
+            y = radius * np.sin(np.radians(360 / pmts_in_row * i))
+            z = 0.0
+
+            # If the PMT is outside of the pit move it up.
+            if radius > watertank.tank_pit_radius:
+                z = watertank.tank_pit_height
+
+            channelmap[name] = copy.deepcopy(configs["pmts"])
+            channelmap[name]["daq"]["rawid"] = rawid
+            rawid += 1
+            channelmap[name]["name"] = name
+            channelmap[name]["location"] = {"name": "floor", "x": x, "y": y, "z": z}
+            channelmap[name]["location"]["direction"] = {"nx": 0, "ny": 0, "nz": 1}
+
+    # The wall PMTs require some polygon math
+    faces = configs["pmts_pos"]["tyvek"]["faces"]
+    # Geant4 uses r as inscribe radius, but we need the circumradius
+    radius = configs["pmts_pos"]["tyvek"]["r"] / np.cos(np.pi / faces)
+
+    # Compute vertices of the polygon
+    vertices = [
+        (radius * np.cos(2 * np.pi * i / faces), radius * np.sin(2 * np.pi * i / faces)) for i in range(faces)
+    ]
+    for row in configs["pmts_pos"]["wall"].values():
+        row_index = row["id"]
+        pmts_in_row = row["n"]
+        z = row["z"]
+
+        # Distribute detectors evenly across faces
+        detectors_per_face = pmts_in_row // faces  # How many detectors per face (integer division)
+        extra_detectors = pmts_in_row % faces  # Remaining detectors to distribute
+        pmt_id = 0
+
+        # Now some crazy algorithm to distribute the extra detectors homogeneously
+        # Invented by Lorenz Gebler
+        m = extra_detectors  # short variable names to make the code more readable
+        n = faces
+        # Try splitting the polygon faces in repetitive cells
+        scl = n // m  # shortest cell length
+        sc = [0] * scl  # shortest cell
+        sc[0] = 1  # Set the first element to 1
+        extra_detectors_per_face = sc * m
+        # In case we cannot split the polygon in equal cells
+        if n % m != 0:
+            k = n - len(extra_detectors_per_face)
+            sclk = m // k
+            sck = sc * sclk + [0]
+            extra_detectors_per_face = sck * k + sc * (m - k)
+        # We need to truncate the list as somehow it creates too big cells
+        extra_detectors_per_face = extra_detectors_per_face[:n]
+
+        for i in range(faces):
+            x1, y1 = vertices[i]
+            x2, y2 = vertices[(i + 1) % faces]  # Wrap around
+
+            # Compute face normal for PMT orientation
+            edge_x = x2 - x1
+            edge_y = y2 - y1
+            normal_x = edge_y
+            normal_y = -edge_x
+
+            # Normalize the normal vector
+            norm_length = np.sqrt(normal_x**2 + normal_y**2)
+            normal_x /= norm_length
+            normal_y /= norm_length
+            normal_z = 0
+
+            # Compute the number of detectors on this face, permutate the extras by the row index
+            num_detectors_this_face = detectors_per_face + extra_detectors_per_face[(i + row_index) % faces]
+
+            for j in range(num_detectors_this_face):
+                name = f"PMT{row_index + 10}{pmt_id + 1:02d}"
+                pmt_id += 1
+                # Interpolate position along the face
+                t = (j + 1) / (num_detectors_this_face + 1)  # Normalized position (avoid exact endpoints)
+                x = x1 * (1 - t) + x2 * t
+                y = y1 * (1 - t) + y2 * t
+
+                channelmap[name] = copy.deepcopy(configs["pmts"])
+                channelmap[name]["daq"]["rawid"] = rawid
+                rawid += 1
+                channelmap[name]["name"] = name
+                channelmap[name]["location"] = {"name": "wall", "x": x, "y": y, "z": z}
+                channelmap[name]["location"]["direction"] = {"nx": normal_x, "ny": normal_y, "nz": normal_z}
+
+        # Check that all PMTs are placed. We do not totally trust the distribution algorithm
+        if pmt_id != pmts_in_row:
+            msg = (
+                "Not all PMTs were placed. Check the distribution algorithm. PMTs placed: "
+                + str(pmt_id)
+                + " PMTs to place: "
+                + str(pmts_in_row)
+            )
+            raise ValueError(msg)
+
+
+def generate_special_metadata(string_idx: list, hpge_names: list, configs: dict) -> dict:
+    """Generate special_metadata.yaml file."""
+
+    special_output = {}
+
+    special_output["hpge_string"] = {
+        f"{string_idx[i][j] + 1}": {
+            "center": {
+                "x_in_mm": configs["array"]["center"]["x_in_mm"][i],
+                "y_in_mm": configs["array"]["center"]["y_in_mm"][i],
+            },
+            "angle_in_deg": configs["array"]["angle_in_deg"][j],
+            "radius_in_mm": configs["array"]["radius_in_mm"],
+            "rod_radius_in_mm": configs["string"]["copper_rods"]["r_offset_from_center"],
+        }
+        for i, j in np.ndindex(string_idx.shape)
+    }
+
+    special_output["hpges"] = {
+        f"{name}": {"rodlength_in_mm": configs["string"]["units"]["l"], "baseplate": "xlarge"}
+        for name in hpge_names
+    }
+
+    special_output["fibers"] = {
+        f"S{string + 1:02d}{n + 1:02d}": {
+            "name": f"S{string + 1:02d}{n + 1:02d}",
+            "type": "single_string",
+            "geometry": {"tpb": {"thickness_in_nm": 1093}},
+            "location": {
+                "x": float(
+                    configs["array"]["center"]["x_in_mm"][string // len(configs["array"]["angle_in_deg"])]
+                    + configs["array"]["radius_in_mm"]
+                    * np.cos(
+                        np.radians(
+                            configs["array"]["angle_in_deg"][string % len(configs["array"]["angle_in_deg"])]
+                        )
+                    )
+                ),
+                "y": float(
+                    configs["array"]["center"]["y_in_mm"][string // len(configs["array"]["angle_in_deg"])]
+                    - configs["array"]["radius_in_mm"]
+                    * np.sin(
+                        np.radians(
+                            configs["array"]["angle_in_deg"][string % len(configs["array"]["angle_in_deg"])]
+                        )
+                    )
+                ),
+                "module_num": n,
+            },
+        }
+        for string in string_idx.flatten()
+        for n in range(configs["string"]["n_sipm_modules_per_string"])
+    }
+
+    special_output["calibration"] = {}
+
+    special_output["watertank_instrumentation"] = {
+        "tyvek": {
+            "r": configs["pmts_pos"]["tyvek"]["r"],
+            "faces": configs["pmts_pos"]["tyvek"]["faces"],
+        },
+    }
+
+    special_output["detail"] = configs["detail"]
+
+    return special_output
+
+
+def generate_channelmap(
+    string_idx: list,
+    hpge_names: list,
+    hpge_rawid: list,
+    configs: dict,
+    unit_divisor: int = 100,
+) -> dict:
+    """Generate channelmap.json file."""
+
+    channelmap = {}
+
+    crystal_records = normalize_records(configs.get("crystal"))
+    hpge_records = normalize_records(configs.get("hpge"))
+
+    n_diods = len(hpge_records)
+
+    for index, (name, rawid) in enumerate(zip(hpge_names, hpge_rawid, strict=True)):
+        channelmap[name] = copy.deepcopy(hpge_records[index % n_diods])
+        channelmap[name]["name"] = name
+        channelmap[name]["daq"]["rawid"] = rawid
+        channelmap[name]["location"]["string"] = rawid // unit_divisor
+        channelmap[name]["location"]["position"] = rawid % unit_divisor
+        crystal = crystal_records[index % len(crystal_records)]
+        channelmap[name]["production"]["crystal"] = crystal["name"]
+        channelmap[name]["production"]["order"] = crystal["order"]
+
+    max_hpge_rawid = int(max(hpge_rawid)) if len(hpge_rawid) > 0 else 0
+    rawid = sipm_rawid_start = max(5000, _round_up(max_hpge_rawid + 1, 1000))
+    for string in string_idx.flatten():
+        for n in range(configs["string"]["n_sipm_modules_per_string"]):
+            name = f"S{string + 1:02d}{n + 1:02d}T"
+            channelmap[name] = copy.deepcopy(configs["sipm"])
+            channelmap[name]["name"] = name
+            channelmap[name]["location"]["fiber"] = name[:-1]
+            channelmap[name]["location"]["position"] = "top"
+            channelmap[name]["location"]["barrel"] = string + 1
+            channelmap[name]["daq"]["rawid"] = rawid
+            rawid += 1
+
+        for n in range(configs["string"]["n_sipm_modules_per_string"]):
+            name = f"S{string + 1:02d}{n + 1:02d}B"
+            channelmap[name] = copy.deepcopy(configs["sipm"])
+            channelmap[name]["name"] = name
+            channelmap[name]["location"]["fiber"] = name[:-1]
+            channelmap[name]["location"]["position"] = "bottom"
+            channelmap[name]["location"]["barrel"] = string + 1
+            channelmap[name]["daq"]["rawid"] = rawid
+            rawid += 1
+
+    pmt_rawid_start = max(6000, _round_up(rawid, 1000))
+    if sipm_rawid_start != 5000 or pmt_rawid_start != 6000:
+        log.info(
+            "array too large for the default raw ID blocks, using %d for SiPMs and %d for PMTs",
+            sipm_rawid_start,
+            pmt_rawid_start,
+        )
+    calculate_and_place_pmts(channelmap, configs, rawid_start=pmt_rawid_start)
+
+    return channelmap
+
+
+def _round_up(value: int, multiple: int) -> int:
+    """Round ``value`` up to the next integer multiple of ``multiple``."""
+    return (value // multiple) * multiple
+
+
+def compile_pre_compiled_config(configs: dict) -> tuple[dict, dict]:
+    """Compile the pre-compiled configs into the objects the geometry is built from.
+
+    Parameters
+    ----------
+    configs
+        the pre-compiled configs, keyed by file name without its extension. Use
+        :func:`pygeoml1000.config.load_pre_compiled_config` to obtain them.
+
+    Returns
+    -------
+    tuple
+        ``(channelmap, special_metadata)``
+    """
+    string_idx = np.arange(
+        len(configs["array"]["center"]["x_in_mm"]) * len(configs["array"]["angle_in_deg"])
+    ).reshape(len(configs["array"]["center"]["x_in_mm"]), len(configs["array"]["angle_in_deg"]))
+
+    n_units = configs["string"]["units"]["n"]
+
+    unit_width = max(2, len(str(n_units)))
+    string_width = 5 - unit_width
+
+    hpge_names, hpge_rawid = [], []
+    for i in range(string_idx.size):
+        for j in range(n_units):
+            hpge_names.append(f"V{i + 1:0{string_width}d}{j + 1:0{unit_width}d}A")
+            hpge_rawid.append((i + 1) * 10**unit_width + j + 1)
+    hpge_names = np.array(hpge_names)
+    hpge_rawid = np.array(hpge_rawid)
+
+    special_metadata = generate_special_metadata(string_idx, hpge_names, configs)
+    channelmap = generate_channelmap(string_idx, hpge_names, hpge_rawid, configs, unit_divisor=10**unit_width)
+
+    return convert_to_plain_types(channelmap), convert_to_plain_types(special_metadata)
