+++ tests/test_cli.py
@@ -23,13 +23,16 @@ def test_write_config_only(tmp_path):
     assert config.load_config(out) == config.resolve_config({})
 
 
-def test_dump_raw_configs(tmp_path):
+def test_dump_pre_compiled_configs(tmp_path):
     from pygeoml1000 import cli, config
 
-    cli.dump_gdml_cli(["--dump-raw-configs", str(tmp_path)])
+    cli.dump_gdml_cli(["--dump-pre-compiled-configs", str(tmp_path)])
 
     assert (tmp_path / "configs" / "string.yaml").exists()
-    assert utils.load_dict(str(tmp_path / "configs" / "string.yaml")) == config.load_raw_config()["string"]
+    assert (
+        utils.load_dict(str(tmp_path / "configs" / "string.yaml"))
+        == config.load_pre_compiled_config()["string"]
+    )
 
 
 def test_cli_args_override_the_config_file(tmp_path):
@@ -56,7 +59,7 @@ def test_build_from_config_file(tmp_path):
     config_file = tmp_path / "geom-config.yaml"
     gdml_file = tmp_path / "l1000.gdml"
     utils.write_dict(
-        {"assemblies": ["cryostat", "HPGe_dets"], "raw_config": {"string": {"units": {"n": 1}}}},
+        {"assemblies": ["cryostat", "HPGe_dets"], "pre_compiled_config": {"string": {"units": {"n": 1}}}},
         str(config_file),
     )
 
@@ -79,14 +82,14 @@ def test_write_metadata_only(tmp_path):
     assert any(name.startswith(metadata.DIODES_DIR) for name in tree)
 
 
-def test_write_metadata_honours_the_raw_config(tmp_path):
+def test_write_metadata_honours_the_pre_compiled_config(tmp_path):
     """The crystal catalog must come from the config, not from the packaged defaults."""
     from pygeoml1000 import cli, metadata
 
     config_file = tmp_path / "geom-config.yaml"
     out = tmp_path / "metadata.tar.gz"
     utils.write_dict(
-        {"raw_config": {"crystal": [{"name": "042", "order": 7, "slices": {"A": {}}}]}},
+        {"pre_compiled_config": {"crystal": [{"name": "042", "order": 7, "slices": {"A": {}}}]}},
         str(config_file),
     )
